requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
streamlit
groq
openai